First implementation of Vitruvius
Based on a script I'd previously written. This is built with a modular
architecture to allow parsers to be added for other schema types. The
gem ships with a parser for [JSON Schema](http://json-schema.org/).

The `App` class takes a parser and uses it to generate a Sinatra
application. The parser wraps all the API schema specific information
and returns a set of `Endpoint` objects that that the `App` can use to
build the application.

A sample body is built using a schema-specific class that recurses
through the schema. Because it expects key-value pairs, a dummy key is
added for arrays and later stripped. This is a hack.